restructure regex tests
